test: move RepositoryTest into Integration-Test with testcontainer
