+++ build.gradle.kts
@@ -24,25 +24,25 @@ repositories {
 }
 
 sourceSets {
-    create("intTest") {
+    create("integrationTest") {
         compileClasspath += sourceSets.main.get().output
         runtimeClasspath += sourceSets.main.get().output
     }
 }
 
-val intTestImplementation by configurations.getting {
+val integrationTestImplementation by configurations.getting {
     extendsFrom(configurations.implementation.get())
 }
 
-configurations["intTestRuntimeOnly"].extendsFrom(configurations.runtimeOnly.get())
-configurations["intTestImplementation"].extendsFrom(configurations.testImplementation.get())
+configurations["integrationTestRuntimeOnly"].extendsFrom(configurations.runtimeOnly.get())
+configurations["integrationTestImplementation"].extendsFrom(configurations.testImplementation.get())
 
 val integrationTest = task<Test>("integrationTest") {
     description = "Runs integration tests."
     group = "verification"
 
-    testClassesDirs = sourceSets["intTest"].output.classesDirs
-    classpath = sourceSets["intTest"].runtimeClasspath
+    testClassesDirs = sourceSets["integrationTest"].output.classesDirs
+    classpath = sourceSets["integrationTest"].runtimeClasspath
     shouldRunAfter("test")
 }
 
@@ -58,12 +58,11 @@ dependencies {
     runtimeOnly("com.mysql:mysql-connector-j")
     testImplementation("org.springframework.boot:spring-boot-starter-test")
     implementation("org.modelmapper:modelmapper:2.4.2")
-    intTestImplementation("org.springframework.boot:spring-boot-testcontainers")
-    intTestImplementation("org.testcontainers:junit-jupiter")
-    intTestImplementation("org.testcontainers:mysql")
+    integrationTestImplementation("org.springframework.boot:spring-boot-testcontainers")
+    integrationTestImplementation("org.testcontainers:junit-jupiter")
+    integrationTestImplementation("org.testcontainers:mysql")
     implementation("com.ninja-squad:springmockk:4.0.2")
     testImplementation("com.h2database:h2")
-    implementation("org.modelmapper:modelmapper:2.4.2")
 }
 
 tasks.withType<KotlinCompile> {
@@ -122,7 +121,7 @@ tasks.jacocoTestReport {
 
 tasks.jacocoTestCoverageVerification {
 
-    var Qdomains = mutableListOf<String>()
+    val Qdomains = mutableListOf<String>()
 
     for (qPattern in 'A'..'Z') {
         Qdomains.add("*.Q$qPattern*")
