--- src/intTest/resources/application.properties
@@ -1,2 +1,0 @@
-spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.ddl-auto=none
