+++ src/integrationTest/kotlin/com/group4/ticketingservice/TestControllerTest.kt
@@ -1,0 +1,23 @@
+package com.group4.ticketingservice
+
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
+
+@AutoConfigureMockMvc
+class TestControllerTest : AbstractIntegrationTest() {
+    @Autowired private lateinit var mockMvc: MockMvc
+
+    @Test
+    fun contextLoads() {
+    }
+
+    @Test
+    fun `E2E test`() {
+        mockMvc.perform(MockMvcRequestBuilders.get("/bookmarks"))
+            .andExpect(status().isOk)
+    }
+}
