+++ src/integrationTest/resources/application.properties
@@ -1,0 +1,1 @@
+spring.jpa.hibernate.ddl-auto=create
