+++ src/integrationTest/resources/db/mysql/conf.d/my.cnf
