+++ src/integrationTest/resources/db/mysql/initdb.d/init-db.sql
@@ -1,0 +1,1 @@
+-- use ddl-auto
