--- src/main/kotlin/com/group4/ticketingservice/dto/UserFromdto.kt
@@ -1,6 +1,0 @@
-package com.group4.ticketingservice.dto
-
-data class UserFromdto(
-    var user_id: Int,
-    var show_id: Int
-)
