--- src/main/kotlin/com/group4/ticketingservice/entity/User.kt
@@ -1,15 +1,0 @@
-package com.group4.ticketingservice.entity
-
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-
-@Entity
-class User(
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Int? = null,
-    var user_id: Int, // 사용자 식별자
-    var show_id: Int // 공연 식별자
-)
