--- src/main/kotlin/com/group4/ticketingservice/repository/TestRepository.kt
@@ -1,8 +1,0 @@
-package com.group4.ticketingservice.repository
-
-import com.group4.ticketingservice.entity.User
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface TestRepository : JpaRepository<User, Long>
