--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.datasource.url=jdbc:mysql://${MYSQL_HOST}:${MYSQL_PORT}/${MYSQL_SCHEMA}
-spring.datasource.username=${MYSQL_USERNAME}
-spring.datasource.password=${MYSQL_PASSWORD}
-spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.ddl-auto=validate
