--- src/test/kotlin/com/group4/ticketingservice/TicketingserviceApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.group4.ticketingservice
-
-import org.junit.jupiter.api.Test
-
-// @SpringBootTest
-class TicketingserviceApplicationTests {
-
-    @Test
-    fun contextLoads(): Boolean {
-        return true
-    }
-}
